created a new directory for Appo.js
