+++ src/Appo/index.js
@@ -1,0 +1,3 @@
+import Appo from './Appo.js';
+
+export default Appo;
